+++ recipe/meta.yaml
@@ -43,9 +43,6 @@ requirements:
     - ai-edge-litert ==1.3.0
 
 test:
-  files:
-    - tests/keras_backend_test.py
-    - run_isolated_test.sh
   source_files:
     - pyproject.toml
     - integration_tests
@@ -61,9 +58,6 @@ test:
     - pytest
   commands:
     - pip check
-    - bash run_isolated_test.sh jax
-    - bash run_isolated_test.sh torch
-    - bash run_isolated_test.sh tensorflow
 
 about:
   home: https://github.com/keras-team/keras
